Add the base code I prepared
This is based on the base code from supernapie.github.io

Albeit slightly modified, css and svg will not be loaded from network. And I will only use system fonts.

If this initial commit message is still visible on one of the files at
the end of the jam, the base code is in good shape.

However I already know I dont't like the plan I had for
initializing the state machine.